Generalize geometry parameters
Do not hardcode a quadrotor geometry but use a generalized `G' matrix instead,
mapping from squared propeller velocities to wrench.  Also add a tunable
inertia matrix, that is not used yet. The `G' matrix defaults to a mikrokopter
quadrotor with the Kf of the 10'' propellers.

This changes the set_geom() setter: it now need the G matrix (6x8 propellers
max) and the inertia matrix.

The controller now outputs directly the propeller velocities, including thrust
and torque limitations. The geometry computations are not ad-hoc anymore.

While here, rename set_vlimit() to set_wlimit(), as propeller velocities are
usually named w.