Add compliance through an admittance controller
The compliance can be turned off an on via set_af_enable().
Parameters (damping, stiffness, inertia) can be set with sef_af_parameters().
Desired external wrench is set with set_af_wrench().

This controller might be moved later to a separate component, but it is added
here for initial testing/debugging.