Add an emergency state in the permanent activity automata
In case of invalid (or absent) state, switch to the emergency descent
state. This does not change the previous behaviour, but adds an explicit state
in the permanent activity automata. This addresses issue #326.

The code checking the validity of the state is factored out in a separate
function so that its return code can be used to switch to the appropriate
state.